feat: 프로젝트 공고 페이지네이션 기능 개발

+++ src/main/java/com/gongjakso/server/domain/post/service/PostService.java
@@ -1,53 +1,76 @@
 package com.gongjakso.server.domain.post.service;
 
 import com.gongjakso.server.domain.member.entity.Member;
-import com.gongjakso.server.domain.member.repository.MemberRepository;
+import com.gongjakso.server.domain.post.common.Pagination;
+import com.gongjakso.server.domain.post.dto.GetProjectRes;
 import com.gongjakso.server.domain.post.dto.PostDeleteRes;
 import com.gongjakso.server.domain.post.dto.PostReq;
 import com.gongjakso.server.domain.post.dto.PostRes;
 import com.gongjakso.server.domain.post.entity.Post;
+import com.gongjakso.server.domain.post.entity.PostScrap;
+import com.gongjakso.server.domain.post.entity.StackName;
+import com.gongjakso.server.domain.post.enumerate.StackNameType;
 import com.gongjakso.server.domain.post.repository.PostRepository;
+import com.gongjakso.server.domain.post.repository.PostScrapRepository;
 import com.gongjakso.server.global.exception.ApplicationException;
-import io.github.classgraph.PackageInfo;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.gongjakso.server.global.exception.ErrorCode.NOT_FOUND_EXCEPTION;
-import static com.gongjakso.server.global.exception.ErrorCode.UNAUTHORIZED_EXCEPTION;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.gongjakso.server.domain.post.enumerate.PostStatus.RECRUITING;
+import static com.gongjakso.server.global.exception.ErrorCode.*;
+
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class PostService {
-
     private final PostRepository postRepository;
+    private final PostScrapRepository postScrapRepository;
+
 
     @Transactional
     public PostRes create(Member member, PostReq req) {
-        Post post = new Post(member, req);
-        postRepository.save(post);
+        Post entity = new Post(req.getTitle(), member, req.getContents(), req.getStatus(), req.getStartDate(), req.getEndDate(),
+                req.getFinishDate(), req.getMaxPerson(), req.getMeetingMethod(), req.getMeetingArea(), req.isQuestionMethod(),
+                req.getQuestionLink(), req.isPostType(), new ArrayList<>());
+
+        List<StackName> stackNames = req.getStackNames().stream()
+                .map(stackNameReq ->  new StackName(entity, stackNameReq.getStackNameType().toString(), stackNameReq.getSize()))
+                .collect(Collectors.toList());
+        entity.getStackNames().addAll(stackNames);
+
+        postRepository.save(entity);
 
         return PostRes.builder()
-                .postId(post.getPostId())
-                .memberId(post.getMember().getMemberId())
-                .title(post.getTitle())
-                .contents(post.getContents())
-                .status(post.getStatus())
-                .startDate(post.getStartDate())
-                .endDate(post.getEndDate())
-                .finishDate(post.getFinishDate())
-                .maxPerson(post.getMaxPerson())
-                .meetingMethod(post.getMeetingMethod())
-                .meetingArea(post.getMeetingArea())
-                .questionMethod(post.isQuestionMethod())
-                .questionLink(post.getQuestionLink())
-                .postType(post.isPostType())
-                .createdAt(post.getCreatedAt())
-                .modifiedAt(post.getModifiedAt())
-                .deletedAt(post.getDeletedAt())
+                .postId(entity.getPostId())
+                .memberId(entity.getMember().getMemberId())
+                .title(entity.getTitle())
+                .contents(entity.getContents())
+                .status(entity.getStatus())
+                .startDate(entity.getStartDate())
+                .endDate(entity.getEndDate())
+                .finishDate(entity.getFinishDate())
+                .maxPerson(entity.getMaxPerson())
+                .stackNames(entity.getStackNames())
+                .meetingMethod(entity.getMeetingMethod())
+                .meetingArea(entity.getMeetingArea())
+                .questionMethod(entity.isQuestionMethod())
+                .questionLink(entity.getQuestionLink())
+                .postType(entity.isPostType())
+                .createdAt(entity.getCreatedAt())
+                .modifiedAt(entity.getModifiedAt())
+                .deletedAt(entity.getDeletedAt())
                 .build();
+
     }
 
     @Transactional
@@ -87,7 +110,12 @@ public class PostService {
             throw new ApplicationException(UNAUTHORIZED_EXCEPTION);
         }
 
-        entity.modify(req);
+        entity.getStackNames().clear();
+
+        List<StackName> updatedStackNames = req.getStackNames().stream()
+                .map(stackNameReq ->  new StackName(entity, stackNameReq.getStackNameType().toString(), stackNameReq.getSize()))
+                .collect(Collectors.toList());
+        entity.getStackNames().addAll(updatedStackNames);
 
         return PostRes.builder()
                 .postId(entity.getPostId())
@@ -99,6 +127,7 @@ public class PostService {
                 .endDate(entity.getEndDate())
                 .finishDate(entity.getFinishDate())
                 .maxPerson(entity.getMaxPerson())
+                .stackNames(entity.getStackNames())
                 .meetingMethod(entity.getMeetingMethod())
                 .meetingArea(entity.getMeetingArea())
                 .questionMethod(entity.isQuestionMethod())
@@ -108,12 +137,17 @@ public class PostService {
                 .modifiedAt(entity.getModifiedAt())
                 .deletedAt(entity.getDeletedAt())
                 .build();
+
     }
 
     @Transactional
     public PostDeleteRes delete(Member member, Long id) {
         Post entity = postRepository.findByPostIdAndDeletedAtIsNull(id)
                 .orElseThrow(() -> new ApplicationException(NOT_FOUND_EXCEPTION));
+
+        if(!member.getMemberId().equals(entity.getMember().getMemberId())){
+            throw new ApplicationException(UNAUTHORIZED_EXCEPTION);
+        }
 
         postRepository.delete(entity);
         return PostDeleteRes.builder()
@@ -123,15 +157,143 @@ public class PostService {
     }
 
     /*
-    public Page<GetProjectRes> getProjects(Pageable page) throws ApplicationException {
+    전체 프로젝트 공고 목록 조회
+     */
+    public Page<GetProjectRes> getProjects(String sort, Pageable p) throws ApplicationException {
+        int page = p.getPageNumber();
+        int size = p.getPageSize();
         try {
-            return postRepository.findAll(page).map(projects -> new GetProjectRes(
-                    projects.getPostId(),
-                    projects.getTitle(),
-                    projects.getMeetingArea());
+            Pagination pagination = new Pagination((int) postRepository.count(), page, size);
+            Pageable pageable = PageRequest.of(pagination.getPage(), size);
+            Page<Post> posts;
+            if(sort.equals("createdAt,desc")){ //최신순
+                posts = postRepository.findAllByPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusOrderByCreatedAtDesc(LocalDateTime.now(), RECRUITING, pageable);
+            } else{ //스크랩순
+                posts = postRepository.findAllByPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusOrderByScrapCountDescCreatedAtDesc(LocalDateTime.now(), RECRUITING, pageable);
+            }
+            return posts.map(post -> new GetProjectRes(
+                    post.getPostId(),
+                    post.getTitle(),
+                    post.getMember().getName(),
+                    post.getStatus(),
+                    post.getStartDate(),
+                    post.getFinishDate()
+            ));
         } catch (Exception e) {
             throw new ApplicationException(INVALID_VALUE_EXCEPTION);
         }
     }
-    */
+
+    /*
+    검색어 기반 프로젝트 공고 목록 조회
+     */
+    public Page<GetProjectRes> getProjectsBySearchWord(String sort, String searchWord, Pageable page) throws ApplicationException {
+        try {
+            Pagination pagination = new Pagination((int) postRepository.count(), page.getPageNumber(), page.getPageSize());
+            Pageable pageable = PageRequest.of(pagination.getPage(), page.getPageSize());
+            searchWord = searchWord.replaceAll(" ", ""); // 검색어에서 공백 제거
+            Page<Post> posts;
+            if (sort.equals("createdAt,desc")) {
+                posts = postRepository.findAllByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusOrderByCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, pageable);
+            } else{
+                posts = postRepository.findAllByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusOrderByScrapCountDescCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, pageable);
+            }
+            return posts.map(post -> new GetProjectRes(
+                    post.getPostId(),
+                    post.getTitle(),
+                    post.getMember().getName(),
+                    post.getStatus(),
+                    post.getStartDate(),
+                    post.getFinishDate()
+            ));
+        } catch (Exception e) {
+            throw new ApplicationException(INVALID_VALUE_EXCEPTION);
+        }
+    }
+
+    /*
+    지역, 스택 기반 프로젝트 공고 목록 조회
+     */
+    public Page<GetProjectRes> getProjectsByMeetingAreaAndStackNameAndSearchWord(
+            String sort, String meetingArea, String stackName, String searchWord, Pageable page) throws ApplicationException {
+        try {
+            Pagination pagination = new Pagination((int) postRepository.count(), page.getPageNumber(), page.getPageSize());
+            Pageable pageable = PageRequest.of(pagination.getPage(), page.getPageSize());
+            searchWord = searchWord.replaceAll(" ", "");
+            if(!stackName.isBlank()) {
+                if (!StackNameType.isValid(stackName)){
+                    throw new ApplicationException(INVALID_VALUE_EXCEPTION);
+                }
+                Page<Post> posts;
+                if (sort.equals("createdAt,desc")) {
+                    posts = postRepository.findAllPostsJoinedWithStackNamesByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusAndMeetingAreaContainsAndStackNamesStackNameTypeContainsOrderByCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, meetingArea, stackName.toString(), pageable);
+                }else{
+                    posts = postRepository.findAllPostsJoinedWithStackNamesByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusAndMeetingAreaContainsAndStackNamesStackNameTypeContainsOrderByScrapCountDescCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, meetingArea, stackName.toString(), pageable);
+                }
+                return posts.map(post -> new GetProjectRes(
+                    post.getPostId(),
+                    post.getTitle(),
+                    post.getMember().getName(),
+                    post.getStatus(),
+                    post.getStartDate(),
+                    post.getFinishDate()
+                ));
+            } else{
+                Page<Post> posts;
+                if (sort.equals("createdAt,desc")) {
+                    posts = postRepository.findAllByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusAndMeetingAreaContainsOrderByCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, meetingArea, pageable);
+                }else{
+                    posts = postRepository.findAllByTitleContainsAndPostTypeTrueAndDeletedAtIsNullAndFinishDateAfterAndStatusAndMeetingAreaContainsOrderByScrapCountDescCreatedAtDesc(searchWord.toLowerCase(), LocalDateTime.now(), RECRUITING, meetingArea, pageable);
+                }
+                return posts.map(post -> new GetProjectRes(
+                        post.getPostId(),
+                        post.getTitle(),
+                        post.getMember().getName(),
+                        post.getStatus(),
+                        post.getStartDate(),
+                        post.getFinishDate()
+                ));
+            }
+        } catch (Exception e) {
+            throw new ApplicationException(INVALID_VALUE_EXCEPTION);
+        }
+    }
+
+    /*
+        공고 스크랩 기능
+     */
+    @Transactional
+    public void scrapPost(Member member, Long postId) {
+        try {
+            Post post = postRepository.findByPostIdAndDeletedAtIsNull(postId)
+                    .orElseThrow(() -> new ApplicationException(NOT_FOUND_EXCEPTION));
+            if (member.getMemberId() == null) {
+                throw new ApplicationException(UNAUTHORIZED_EXCEPTION);
+            }
+
+            PostScrap postScrap = postScrapRepository.findByPostAndMember(post, member);
+
+            if (postScrap == null) { //첫 스크랩
+                postScrap = PostScrap.builder()
+                        .post(post)
+                        .member(member)
+                        .scrapStatus(true)
+                        .build();
+                post.setScrapCount(post.getScrapCount() + 1);
+            } else { // 스크랩 한 적 있는 경우
+                if (postScrap.getScrapStatus() == true) { //스크랩한 상태면 취소
+                    postScrap.setScrapStatus(false);
+                    if (post.getScrapCount() > 0) post.setScrapCount(post.getScrapCount() - 1);
+                    else throw new ApplicationException(NOT_FOUND_EXCEPTION);
+                } else { //스크랩 안 한 경우
+                    postScrap.setScrapStatus(true);
+                    post.setScrapCount(post.getScrapCount() + 1);
+                }
+            }
+            postScrapRepository.save(postScrap);
+            postRepository.save(post);
+        }catch(Exception e){
+            throw new ApplicationException(NOT_FOUND_EXCEPTION);
+        }
+    }
 }
